+++ tools/cli/commands.js
@@ -82,9 +82,9 @@ export function parseServerOptionsForRunCommand(options, runTargets) {
       isRunOnDeviceRequested);
   }
 
-  const parsedLocalServerPort = parseLocalServerPortOption(options);
+  const parsedCordovaServerPort = parseCordovaServerPortOption(options);
 
-  return { parsedServerUrl, parsedMobileServerUrl, parsedLocalServerPort };
+  return { parsedServerUrl, parsedMobileServerUrl, parsedCordovaServerPort };
 }
 
 function parsePortOption(portOption) {
@@ -112,9 +112,9 @@ function parseMobileServerOption(mobileServerOption,
   return parsedMobileServerUrl;
 }
 
-function parseLocalServerPortOption(options = {}) {
-  const localServerPortOption = options['local-server-port'];
-  return localServerPortOption ? parseInt(localServerPortOption, 10) : null;
+function parseCordovaServerPortOption(options = {}) {
+  const cordovaServerPortOption = options['cordova-server-port'];
+  return cordovaServerPortOption ? parseInt(cordovaServerPortOption, 10) : null;
 }
 
 function detectMobileServerUrl(parsedServerUrl, isRunOnDeviceRequested) {
@@ -297,7 +297,7 @@ var runCommandOptions = {
   options: {
     port: { type: String, short: "p", default: DEFAULT_PORT },
     'mobile-server': { type: String },
-    'local-server-port': { type: String },
+    'cordova-server-port': { type: String },
     // XXX COMPAT WITH 0.9.2.2
     'mobile-port': { type: String },
     'app-port': { type: String },
@@ -334,7 +334,7 @@ function doRunCommand(options) {
   // Additional args are interpreted as run targets
   const runTargets = parseRunTargets(options.args);
 
-  const { parsedServerUrl, parsedMobileServerUrl, parsedLocalServerPort } =
+  const { parsedServerUrl, parsedMobileServerUrl, parsedCordovaServerPort } =
     parseServerOptionsForRunCommand(options, runTargets);
 
   var includePackages = [];
@@ -410,7 +410,7 @@ function doRunCommand(options) {
         const cordovaProject = new CordovaProject(projectContext, {
           settingsFile: options.settings,
           mobileServerUrl: utils.formatUrl(parsedMobileServerUrl),
-          localServerPort: parsedLocalServerPort });
+          cordovaServerPort: parsedCordovaServerPort });
         if (buildmessage.jobHasMessages()) return;
 
         cordovaRunner = new CordovaRunner(cordovaProject, runTargets);
@@ -440,7 +440,7 @@ function doRunCommand(options) {
     mongoUrl: process.env.MONGO_URL,
     oplogUrl: process.env.MONGO_OPLOG_URL,
     mobileServerUrl: utils.formatUrl(parsedMobileServerUrl),
-    localServerPort: parsedLocalServerPort,
+    cordovaServerPort: parsedCordovaServerPort,
     once: options.once,
     noReleaseCheck: options['no-release-check'] || process.env.METEOR_NO_RELEASE_CHECK,
     cordovaRunner: cordovaRunner
@@ -898,7 +898,7 @@ var buildCommands = {
     "server-only": { type: Boolean },
     'mobile-settings': { type: String },
     server: { type: String },
-    "local-server-port": { type: String },
+    "cordova-server-port": { type: String },
     // XXX COMPAT WITH 0.9.2.2
     "mobile-port": { type: String },
     // Indicates whether these build is running headless, e.g. in a
@@ -997,7 +997,7 @@ var buildCommand = function (options) {
 
   let cordovaPlatforms;
   let parsedMobileServerUrl;
-  let parsedLocalServerPort;
+  let parsedCordovaServerPort;
   if (!serverOnly) {
     cordovaPlatforms = projectContext.platformList.getCordovaPlatforms();
 
@@ -1020,7 +1020,7 @@ on an OS X system.");
       }
       parsedMobileServerUrl = parseMobileServerOption(mobileServerOption,
         'server');
-      parsedLocalServerPort = parseLocalServerPortOption(options);
+      parsedCordovaServerPort = parseCordovaServerPortOption(options);
     }
   } else {
     cordovaPlatforms = [];
@@ -1110,7 +1110,7 @@ ${Console.command("meteor build ../output")}`,
         cordovaProject = new CordovaProject(projectContext, {
           settingsFile: options.settings,
           mobileServerUrl: utils.formatUrl(parsedMobileServerUrl),
-          localServerPort: parsedLocalServerPort });
+          cordovaServerPort: parsedCordovaServerPort });
         if (buildmessage.jobHasMessages()) return;
 
         const pluginVersions = pluginVersionsFromStarManifest(
@@ -1564,7 +1564,7 @@ testCommandOptions = {
   options: {
     port: { type: String, short: "p", default: DEFAULT_PORT },
     'mobile-server': { type: String },
-    'local-server-port': { type: String },
+    'cordova-server-port': { type: String },
     // XXX COMPAT WITH 0.9.2.2
     'mobile-port': { type: String },
     'debug-port': { type: String },
@@ -1663,7 +1663,7 @@ function doTestCommand(options) {
   const runTargets = parseRunTargets(_.intersection(
     Object.keys(options), ['ios', 'ios-device', 'android', 'android-device']));
 
-  const { parsedServerUrl, parsedMobileServerUrl, parsedLocalServerPort } =
+  const { parsedServerUrl, parsedMobileServerUrl, parsedCordovaServerPort } =
     parseServerOptionsForRunCommand(options, runTargets);
 
   // Make a temporary app dir (based on the test runner app). This will be
@@ -1881,7 +1881,7 @@ function doTestCommand(options) {
         const cordovaProject = new CordovaProject(projectContext, {
           settingsFile: options.settings,
           mobileServerUrl: utils.formatUrl(parsedMobileServerUrl),
-          localServerPort: parsedLocalServerPort });
+          cordovaServerPort: parsedCordovaServerPort });
         if (buildmessage.jobHasMessages()) return;
 
         cordovaRunner = new CordovaRunner(cordovaProject, runTargets);
@@ -1900,7 +1900,7 @@ function doTestCommand(options) {
     options,
     {
       mobileServerUrl: utils.formatUrl(parsedMobileServerUrl),
-      localServerPort: parsedLocalServerPort,
+      cordovaServerPort: parsedCordovaServerPort,
       proxyPort: parsedServerUrl.port,
       proxyHost: parsedServerUrl.hostname,
     }